requirements.txt
boto3==1.18.42
botocore==1.21.42
jmespath==0.10.0
paho-mqtt==1.5.1
pymongo==3.12.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
uuid==1.30
